ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-multipart==0.0.6
pillow==10.1.0
numpy
opencv-python-headless
ultralytics
requests
python-dotenv
